+++ src/main/java/roomescape/reservation/ui/dto/CreateReservationTimeResponse.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import roomescape.reservation.domain.ReservationTime;
 
-public record ReservationTimeResponse(
+public record CreateReservationTimeResponse(
         Long id,
         @JsonFormat(pattern = "HH:mm") LocalTime startAt
 ) {
 
-    public static ReservationTimeResponse from(final ReservationTime reservationTime) {
-        return new ReservationTimeResponse(
+    public static CreateReservationTimeResponse from(final ReservationTime reservationTime) {
+        return new CreateReservationTimeResponse(
                 reservationTime.getId(),
                 reservationTime.getStartAt()
         );
