+++ src/test/java/roomescape/reservation/application/ReservationTimeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.service;
+package roomescape.reservation.application;
 
 import java.time.LocalTime;
 import org.assertj.core.api.Assertions;
@@ -12,7 +12,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.exception.AlreadyExistException;
 import roomescape.exception.ResourceNotFoundException;
-import roomescape.reservation.application.ReservationTimeService;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.ReservationTimeRepository;
