+++ src/test/java/roomescape/reservation/infrastructure/JdbcReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.infrastructure.JdbcReservationTimeRepository;
 
 @JdbcTest
 @Import(JdbcReservationTimeRepository.class)
