+++ src/test/java/roomescape/theme/application/ThemeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.theme.service;
+package roomescape.theme.application;
 
 
 import java.util.List;
